168.2 ComponentWillUnmount lifecycle method(Section 21: React Redux Application)
So either you auction block post handload and will dispatch this action to signal to our reduce or that

the data that was in state inside the block post-state is no longer valid so it should be reset.

Now inside reducer will handle this new action by just simply setting is fetching to false and post

to NO and what we do that for is that inside our blog posts contain or.

Well let's first maybe import the action creator to my dispatch two props.

Mind that you need a real import from the model and cold this props block post unload to dispatch is

to reduce her when the component will on the Mount.

So when the component will no longer be in the room it won't be rendered anymore.

And this happens when you will change a route.

You will reset the state of an individual block post so to see that actually.

How that works.

Let's set a timeout for the component did mount and comment out the unload metalled And let's go to

our blog post now.

It's loading after four seconds because we use timeout and when you go to a different blog posts you'll

see that the old title is displayed and only after 4 seconds when we fetch block post it displays the

new title.

So that's why we have this component will unmount and there we are resetting the state so you can see

now it shows that block post does not exist and only after 4 seconds loads the new data because by default

it is no.

So we are not showing an older data.

Each time a component is mounting.

So when you have components like that that change depending on some parameters is a good idea to add

a component will amount to handle such cases so you can reset the state for this particular order used

inside this particular component.