170. Spinner and Message components (loading state, simple message)(Section 21: React Redux Application)
Now inside our blog post The View and also probably later we have two cases where we display the same

thing.

And one is the shoulder loader.

Well in a blog post least we do the same with a shoulder loader when the component is still fetching

the data or we show a simple message that nothing was found.

If we are unable to either download or there is some error or.

Well maybe that might not be a blog post or blog post list is empty.

So we create another component that we can just use and insert here.

And well that does not repeat ourselves with the same code.

And also we can make it nicer.

So let's our new class spinner inside components and we will make it a little bit nicer than it is right

now.

So we'll put it inside a car at the

with the shadows so it will always it always could be used for the places where we use this car from

Bootstrap to this plate either a block post or a blog post list.

So the spinner should also extend our re-act component and just have to render metal.

So it's basically not doing anything.

Doesn't even have any property.

It just shows the spinner.

So now if the data is still fetching inside the block post component we can just show the spinner and

we can do the same inside that block post list.

So when that data is still fetching we are just showing the spinner component that we have just created.

Don't forget our imports statement

now after refreshing you can see that it looks a little bit better because it starts with those div

which is a card and has the shadows so the interface looks more

consistent.

Now we will also create another component called mess search.

And this one will actually have one property.

We can just copy what we have from spinner because this component will be almost exactly the same.

And instead of the spinner inside it will show a message.

So we get it from props that means we'll be also passing it inside a property to the message component.

And we only display this message inside the card body.

So now we can replace this div inside block post and block posts list and show a message component with

a message block post does not exist.

We can also do the same insider blog post least and even later when we have another case like this we

would like to display a generic message or a spinner.

We could then use our components.

Now we can see that it looks pretty nice.

It shows the loader and also a nice message when something bad actually has happened.

Also if we stop our server you can see how it all looks.

We have the spinner and on error we immediately showed the message that something was wrong.

So let's restart the pute server now and you will see the spinner.

It takes a little bit more time when you make a first request after restarting.

Now let's see if it's a good idea to control the fetching frauleins site or displaying the spinner from

inside block posts least component.

As we said we want this to be just a plain component dumb dummy component that just displays data I

shouldn't really care whether the data is fetching or not.

It should just render a block post list.

So yes we will remove that from the block post least because this signals that this control has some

is aware of the state and instead will put this inside block post list container and we won't pass the

fetching property to the block post so we can remove it and also we can remove the import for the spinner

component.

So this makes sense because we wanted this component the block post and block Parsley's to be independent

of the application state.

Now let's do the same thing inside the block post and remove this Pineiro component and only use it

from inside the block post container because it itself is fetching the data so it shouldn't parse this

as a property to the lock post component.

It should just first displayed a spinner and only if it gets the data it should display there either

block post component or the block post least component.