180.1. Token Middleware and JWT authentication plugin 2(Section 21: React Redux Application)
So let's continue with our talk in storage.

When we log in.

So first let's test this.

If our middleware works correctly so we can cancel LUGG from that token middleware when we handle the

user logging success and WorldCom and it's from middleware and send that action token.

Now let's also add to this to the reducer so we can see the order in which all of this happens.

So first let's try to authenticate with some valid username and password.

That could be the admin and the proper admin password.

Well it seems I misspelt their password so let's try that again with the correct one this time.

Well but it seems that how many times we try even with the correct password it doesn't really work.

So we still have 401.

So let's see what's inside it.

So there is inverted JPT talkin and this means that we have this problem because we have the post request

we always try to send a token even or Also even with the GET requests.