180.2. Token Middleware and JWT authentication plugin 2(Section 21: React Redux Application)
So we actually shouldn't do that because before we get a token it's it's basically empty.

So we should add our own couriered function that would first accept argument called secure so this way

we can make sure that we only add that token when we have a valid one so we can only send it to the

resources that really have to be protected because if we try to log in even vote it does not require

to be secured.

But when you send a token and it is no or invalid the API platform will just return at the top consent

was invalid and it won't work.

So we are having a first function that only returns a function that checks whether the token is set.

So by checking if tokened and secure means that security has to be true.

So let's add this parameter to the post and get.

Methods and in get the security would be false by default.

And in post it would be true by default because probably most of our post actions require this parameter.

We try to be secure and get does not.

So let's also add this inside a token plugin.

So we are calling that token Plug-In waive a parameter and only then it returns a function that is the

actual plugin.

So it's also a corporate function.

The same thing we did before now when we added the false parameter for our call to the logging check.

So we didn't want it to be secured.

Now you can see in the console that we have our data from form and the token from middleware.

And only then from the reducer.

So that proves that middleware is called after the action is dispatched but before it gets to that reduce

our.