180.3. Token Middleware and JWT authentication plugin 2(Section 21: React Redux Application)
We can also modify our requests to see what's coming back from the request itself.

So let's go to the agent and J.S. and modified the post with curly braces so it's not just simple code

that returns a default value.

It will first console log it.

So we have to modify the commerce to the semicolon and the first console lock the token

seems we have an error so I probably forget parentheses.

Yes.

It's the end of the super agent post.

So we have to fix that.

This house means the parentheses.

Now let's refresh and try to log in again using some correct credentials.

Could be for the admin.

Now there is another mistake and it is probably due to our change from the Aero function to the function

surrounded with curly braces because if you don't that's surrounded from curly braces It just takes

the first and only expression as the return value.

And when you modify it like that.

So it has a couple of expressions unique to return.

OK so let's get back to what we were supposed to check first loggin again as the user.

So now from logging inside agent you see that token will know.

But after logging again you'll see that the token is already set to the what was previously said so

by outputting the token variable we can see that it is being stored inside the token variable now.

But if you will refresh the page we lose the token.

So the user would have to authenticate again.

So we all solve that in the next lecture.

So the talkin token as you see is inside the local storage.

So we have to somehow read it when our application mounts for the first time.

That is probably when the user is refreshing our page.