179. Storing JWT token in LocalStorage, Token Middleware(Section 21: React Redux Application)
So last time we created a simple reducer to keep the token and user ID inside the state.

But we will also have to keep the token other outside the state because we are not always using the

state to fetch any data.

Sometimes we will probably need to get the token inside the requests to actually add it to our request.

So for this we will define our own middleware.

So let's create middleware J.S. and define our token middle we're like this.

But this requires a little bit of explanation.

So the token middleware is a curried function.

So the couriered function is a function that returns our function.

So you can see that this is that yes 6 definition of it.

So there is a store next and action.

And this would be how it would look like inside.

Well if would be yes.

5 so the older javascript so the middleware receives the store and then it will return a function that

receives a next function.

And this will return another function that receives an action.

So the store is our Rydex store.

The next is the function that our middleware will call when it's like done with anything.

It has to do and it's sense that action to reduce hours or to another middleware and the action is our

current action that is being dispatched.

Now we've read Rick Stengel you can see how this would be actually cold.

So this is how we call the curt function.

So we passed the first value and then it's like another function call where you pass the second volume

and the first function call which you pass the argument to the first function that is them most nested

inside.

Now this is how the media were would work.

So first we have is a store which keeps our application state and from store we dispatch an action and

then comes middleware.

So the action is being dispatched but before it gets to reducer it goes to middleware.

So middleware can decide if should it should go further or whether you can do something along the way

and we will use that to save our token inside our agent.

J.S. we're restored our request making code.

And only then it will get to reducers So we are just joining in the middle of the route using our token

middleware.

So now let's implement our middleware and we do that like it would be reduce or function.

So we are checking.

What is the auction type.

Because the mirror where we'll go through every dispatch action and we only want their user logging

success and probably use a log out when we implement logging out.

So we add a switch and then after each of those we'll call next action.

So it can handle that action further.

Now inside our case we have user logging success.

We actually would like to store somehow into tokens that came through and we will store it inside their

local storage.

So the local storage of the browser is accessible under the window.

And we can call the set item on this local storage to save some value inside it.

So we'll use the JPT token for our token and the user id for the ID of the user that we got from this

auction.

So this auction is being dispatched when the user successfully Loggins.

So we have this data and now we have this inside the local storage.

But now it would be useful to also store it somewhere inside our agent J.S..

So let's define a variable called talken.

And this time we use let not Konst because it will be subject to change.

So it is no initially but we will be modifying it and lets other sets talkin to our requests object

and this accepts one parameter.

So it is new jayvee the token.

And inside this function we set the token variable to equal this new davit token.

Now we will need a plugin for the super agent and plugin is just a function that it can call.

So this plugins the super agent plugins except one parameter which is a request so we can modify the

request beef before it even gets dispatched.

And we will do that by setting a header which is the outer ring zation header the same as we did with

postman and it will accept a mirror and the token that we have stored so not is that it will just send

no token or any value that we have stored when we actually log in and you add that plugging by using

the use method of the super agent.

So there is this chain of methods and you should call use to actually apply this token plugin which

will set the correct header.

Now the set token function is on yoost And while this is correct we have to first call it from inside

our middle word.

So you can call requests set token and this will be the token that we received with that action.

And finally we need a break because we wont be returning immediately.

We want this next action to be able to process after we do our thing.

Now we have created the middle were but we have to yet pass it to the create store function the same

as we did with the phunk middleware.

So this has to be second parameter so each middleware that you register has to be passed to apply middleware.

So we have to import our token middleware from middleware.

J.S. there are a couple of more things we have to do but this will conclude this lecture and we will

continue in the next one.