181.2 Reading token in App component constructor(Section 21: React Redux Application)
But let's try doing this again.

And you can see that each time we try to change our field like loggin field or password it ultimately

does direct because the fields are also inside the state.

So when you modify any form field the state gets changed then it's get mapped to the props of the component.

And this is causing it to redirect to main page because we haven't changed which property has changed.

So you can get into trouble like that when you do not change.

Do not check which property has changed in the component.

Did update.