185.1. Rendering signed-in username in Header(Section 21: React Redux Application)
All right so now let's modify our output reduce our Because we are requesting the user ID.

So when the user profile will be received we should also get our out the user id parameter to this action

and this would be important because inside our producer we would like to check the user id of the profiles

received is different than the user id of currently authenticated user because we would be using the

same code for fetching user profiles to also fetch profiles of other users on the blog.

So this would be this option then when someone writes a blog post or comment you will be able to see

his profile.

So lets add a new field to our state which would be user data and inside the user profile received.

So when we handle this user profile receive action inside the reducer We will check if the state user

ID is exactly the same as the actual user ID.

And also if the state user data is no.

So this would mean this is only the initial request made after someone has logged in or refresh the

page.

So otherwise we won't be setting the user data.

We will keep it the same as it is.

So it just won't change.

Yes and we are doing this as I said because we only wanted to set the user data if there was no data

before and the user id of the authenticated user has actually changed.

Now let's go back to the up component and we should get this user data from prop's because it's part

of the AUF radius or so that our subtree of this state.

So we should automatically have that inside our app component because we are mapping the state to prop's

now inside our hether component.

We display a message hello user.

So we should modify that to display the actual user full name when he is authenticated.

So if the flag is authenticated is true that means we can safely do that.

But let's create separate render method for rendering the actual user.

So we'll get the user data from the props and if it is if it is no then this would mean that we probably

should have to show a spinner component because this would mean that the user has not indicated or has

completed the authentication process but we are in the process of fetching his profile.

So for these cases we will show the spinner that something is happening.

So the work is in progress and only after the user data is available we can show now not the hello user

but the hello actual user name.

So you can replace this part of their JSE X by calling the render user which would return the proper

component in this place.

Now let's test that in the browser.

So try to sign in as admin as we do usually press the logging button and while you see the spinner that

it was fetching the user profile so well maybe doesn't look so well when we use this component probably

we should actually just copy this element with the classes.

So this is a useful component.

When we added below when we display either comments or block posts but maybe it's not very useful for

this so we can just paste it.

It's pretty simple chunk of code.

Now let's try to sign in again after the page is refreshing so well this looked much better than before.