190.3 Displaying errors in redux-form(Section 21: React Redux Application)
So first import class names from the class names module and now create a variable called classes and

coldest class names because it's a function.

So it accepts any number of arguments.

So we'll be adding first argument for control and modifying the class name of all the fields with the

result of this variable.

So this is a this is typical So we've just added a class name now to conditionally render.

Class will only add an Violet when the error is present so if the error is not all or not undefined.

So by adding another argument with the curly braces we can specify that this should be a conditional

class only when certain certain expression is valid.

True or however so we checked up based on the conditions.

So now we've added is invalid to the form and it is bootstrapped class.

So whenever a field is invalid it would be rendered as read.

And this is Fanks to following their submission.

Error with the plain javascript object that has the field name as a key.

Later we'll find how we can convert our API platform responses to automatically generate that for any

form.