192.2. Animations - animated comment adding{yarn add react-transition-group@2.5.0 --save}(Section 21: React Redux Application)
Run : yarn add react-transition-group@2.5.0 --save
this library help us to animate the way the comments are displayed after their creation
Now let's try to animate them.

So this will install the react Transition Group library with the version you see on the screen so be

sure to install it using this version because it's evolving all the time so whenever you start doing

this course it might be there's something newer and doesn't work the way I'm showing it to you.

So we have to learn on the same version.

So that's pretty important.