192.3. Animations - animated comment adding(Section 21: React Redux Application)
The map loop must be inside of TransitionGroup element and any individual element must be inside CSSTransition element in order that this animation to function. key atribute must be in the top element so it must moved to CSSTransition. Animation doesn't work out of the box some css properties must be declarated.
Now, anytime an comment is added to the list of comments or removed this process will be animated.