198.1. Fetching blog post collection on page change(Section 21: React Redux Application)
Now in this lecture we'll actually implement our Puggy nature component to work with the real data.

So clicking on the next or previous pages or appliques or at least on the numbers will actually fetch

that data from the API.

So let's start with what we have inside our blog post least state so we will need the current page to

store it inside the state.

So let's set a default for the current page value of the state to 1 because we start with the first

page and the page count.

Well we have to set that to know because we never know how many pages are there until the request comes

in.

But while the current page could be just one because if there is no page specified We will start with

the first page.

So let's now an action creator and call it block post least said Page.

So will dispatch an action when we want to modify the current page.

So obviously we need action constant for that.

So let's get now to the Konstanz J.S. and here let's add an import statement so I'm doing this for today

ID because well I have this functionality if you don't you will have to this manually.

But I think the visual studio code can also do it.

Now this action will also contain the actual page numbers so it's pretty simple.

It's only the page number to set.

Now let's go back to blog post least reduce our and well we can maybe remove all those Konsole logs

because they're unnecessarily cluttering our console and we already covered that.

So there there's nothing new we don't need to log log all these messages.

Now let's handle this blog post least said Page.

This will just change what one property of the state which is the current page.

So whenever this action is dispatched the current page will be stored inside the state and will be modified.

Now we will add this auction creator inside our blog post the least container.

So that's one of the actions we would like to map from that to our props using gridlocks.

So let's also make sure that you have imported it and the pugging Adra component will have another property

that will be set page.

So it needs to know what it has to do when the page is to be set because it's dummy.

It doesn't care.

It's not loading anything from the back end.

So that's the best way we can do it.

We will just get the actual dispatcher from the props and parse it as a parameter as a prop to the pugging

major component itself.

So since we have the current page we can also get the current page from the props and pass it to the

Packie nature component.

So it's no longer Damini it's not being automatically set it's being set by from.

But well basically from the State of the application and in this case block post least state because

the pugging nater can use any data.

Now inside our pugging later when it on clique Medhat for the cases where the individual page number

is banked clicked.

So we have our set page property that we have passed from the blog post list container and that's how

it should also be called for the Puggy nature that said Page is the Hendler and we are adding function

on Cliq that just calls dispatches this action set page that is being passed by the Propp.

So basically this component doesn't know that it does something that it does as an action for this component.

It is just calling a function so we can see it works perfectly and the page is being updated.

And that's because we automatically re render this component because our props do change so we dispatch

an action set page.

This changes the current page of the state.

So when it gets gets passed to our pogey nature component it's also renderers because it is a prop on

that buggy nater component.

And when a prop changes the component Riera enters so everything works perfectly from top to bottom

and that pugging nature component is not aware of anything.

It doesn't even know that when you click this individual pages that it is dispatching an action it just

wants a function to call.