198.2. Fetching blog post collection on page change(Section 21: React Redux Application)
Now let's modify the block post list fetched to accept a page parameter with the default one and modify

the requests that we make to include that page parameter and we are only doing underscore page because

in the lecture lectures about API platform we modified the deferral name for the page to underscore

page.

So if you haven't done that or you use different parameter name or you haven't changed it at all.

Don't default one is just a page without the underscore.

Now let's say Think how we can handled fetching the block posts again when the page is changing so we

shoot the component did update lifecycle method to the blog post least container.

And now if and obviously it's getting there.

Brief props which are the previous props of this component before the change.

So this signals that something happened and we verify if the current page has changed.

If it if it did we are dispatching the blog post least fetch with the current page.

So.

Well I guess this makes sense so whenever this component is aware of that page change and it is because

we are getting the state we are mapping the whole state from block post list to this component props

and thanks to that each time this date is changing the props change and when prop's changed the component

gets re rendered.

So whenever the page is changed we will fetch the block posts again.

So now let's try to click it.

And you can see that we have different blog posts so it works perfectly.

We can browse through every page of the articles that we have in our API.