199.1. Query parameter based pagination (page in route parameter)(Section 21: React Redux Application)
So previously we implemented the pugging nation.

So whenever you click on any number that signals a page inside our nature component pollinator component

you would see a different results.

Now how we can store that particular page or send a link to someone that he can open a particular page

of the results inside our re-act up because we have pages for the API.

But we do not have any page for our re-act up and now in this lecture we implement that.

So let's start with our block blog post least container and get Querrey Parham page function.

So this will return whatever is inside that prompts page.

So the query parameters that's where they are.

So this prop's much Barnum's page.

This comes from the re-act router.

So whenever you have a parameter for your components and you you are using redox and mapping the state

to prop's you can access that the Querrey parameters your query parameters using prop's March barrooms.

Now let's also modify the component.

Did update the lifecycle method so we can check that the page inside the euro has changed.

So we are just checking previous properties that contains also the page with the current page and if

it did change we were dispatched the block post list set the page so we would just set the page inside

the state because otherwise when the query parameter changes it doesn't mean anything to our up because

it doesn't care about it.

There is no magic that you can slash five or slash page five and this will work immediately.

Obviously we are depending on the state of our up in the query parameter has nothing to do with states

so we have to modify the state in the response to the change of the query parameter and to well we can

do that handle that inside component update because we are mapping the query programs also to do our

props and they are part of the state itself.