199.2. Query parameter based pagination (page in route parameter)(Section 21: React Redux Application)
And now we need change.

Page händler So we'll get his story and blog post list said Page from this props and do this we'll call

blog post list said Page.

So it will dispatch this action and then it will modify their original.

So now let's see our pugging 8 or so we have our set page that works when you click a particular page.

So instead of just calling blog post least set the page or dispatching it as a response to the click.

We'll use our change page metalled that will also modify the world not only dispatch said page but will

also update the user.

Inside the browser.

So we have to of course being that because otherwise the change page would be called as an event handler

and would not have access to this so it would change context.

So it's all this complicated javascript context things.

So let's check it out.

So when we press the button it seems to actually load their proper data and also show the page.

But then it reverted back to another state so that's no good.

And the problem is we haven't defined the page parameter.

And this is how you make an optional parameter by the way using the question mark.

So it does not have to be specified.

Now let's see page one works perfectly.

Page 2 page 10 and it's all inside the you are well.

So if you just click and enter this euro directly you are immediately inside the page then.

And if you go to rehab block so the default page would be one.

And you see Page 1.