200.1. Pagination - previous/next buttons(Section 21: React Redux Application)
So in this lecture we will work on implementing the patterns for the previews and then next page.

So we have the numbers working.

Whenever you click on them it gets you to the current page.

And also we will get the real page count because we now have hardcoded it to 10 and we know from the

response that there is more than 10 pages.

So let's take a look.

We had the first last and next page and this is do we have removed the setting inside the API platform

so you have to have a platform project opened.

And on the lectures about pagination we've added the pagination Parkfield to true.

And this made the square is also a bit faster because it didn't fetch the page count.

So you only got the next page as a link.

So if there was a next page you got a link to the next page and we need to have the page count to implement

the Paganini's properly so just remove that parameter.

And if you made any global settings for the blog post makes sure that the pagination partial setting

is not enabled for block post nor globally.

All right so now we have to think how we can handle this nation when the user is pressing their previous

or the next page.

So the data we have first last and next so it is under Heidrun view and the Hijrah first last and next

are basically links to our API.

But we do not have the numbers that specify what is the actual page.

So for that we will function to our API utilities and their function will be called Heidrun page count

and based on the collections seems every collection we get back from a PR platform has their Hydra meta

data we can we can check what is the page count.

So first we should verify that the Heider view does exist so if we have them at the data because if

you ask for that.

Jason so your accept hether would be application JSM or any other you wouldn't get that back.

So we have to verify that first so it won't break.

And now let's see what we have inside Hydra view.

There is Hydra Cullom last and it's a link to the page so what we can do is we can return a number by

the way by returning number starting with uppercase.

You just typecast the stream to number.

So it's important when you compare varios using triple comparisons like they're free equal signs.

Now we can use regular expressions to fetch the page so under the Hijrah view there is hydrous last

and it contains a page parameter so mind that you use the page parameter name that you use inside API

platform.

And what we expect is the page equals and inside it is our Suppe expression that we will fetch the number

from and slash D-plus means a number that is of land of one or more.

So this would mean that we expect to find a number.

Basically we expect numbers in the stream.

That would be line for one or more.

So whatever it might be 110 hundred and this way will fetch the page count.

So now let's call our Heidrek page count Mattox also mind to imported first insight or blog posts these

three douceur.

And when the blackboard at least received action will be dispatched and we handled up inside the reducer.

We will.

That way we'll be able to fetch the current page count of the elements that we are browsing.

So in our case it's of course the blog posts.

Now you can see that we are also logging it and finally after it is being fetched.

So in the first render we have this page count.

So it says 11.

And it looks correct we've checked that we have 11 pages so now we can pass this to the pugging ATR

component.

Let's also reformat it because the lines get too long.

And now we can see we have 11 pages that we missed before when we hardcoded it to 10 pages.