200.2. Pagination - previous/next buttons(Section 21: React Redux Application)
So what we are left with is the previous and the next the buttons.

So we have a method that handles page changing to the particular page which is the change page mapped

out.

And we also need two more handlers and they would be on the next page click and on previous page Khaliq

So our pugging Natal component will also need to have new props.

This would be to to actually handle those clicks because the pogey Nader component wouldn't be able

to figure out what it should do when those buttons are click because it's not aware of their redock

state and it's not putting any action so it's pure component you can call it a pure component it just

simply ran their stuff doesn't care about anything else.

So how we can calculate their next page so we can use their minimum function from the Moff.

So it's a built in javascript library.

And then new page would be the content page plus one or the page count.

So we are using the minimum.

So for example the current page is 11 and you add one to it and the page count is also 11.

You wouldn't like to move person to 12 page which doesn't exist.

You would like to move the person to 11 so whatever is minimum current page plus one or the page page

count.

And finally we of course called that this change page to actually this at the blog post Lissette page

and modified the rail and on the previous page click we do a different thing.

We are using the maximal function which is taken from the current page minus 1 or 1 so in the same case

if we would be on the first page you wouldn't like to move to the page zero.

You would like to take that bigger number of those too.

This is what the max function does.

It takes a bigger number and it's only useful for the cases where you are on the first page and you

press previous again and well this would move you to the page zero which doesn't exist so obviously

we have to set the page two to be minimum of one.

Now we are passing this metal to our pugging eight components so the next Bache prop and the previous

page Propp and we also called being on those metals because the event handlers and when assigned to

an HD element when they would be cold you wouldn't have access to this.

So they will be called from different contexts so there'll be no days.

It can be undefined or something else so that's why you have to call past them using this.

And the function and the beyond and beyond that this context which in those particular place is actually

our component class.

So that's what you have to do with event handlers.

Now let's add this handlers to our buttons.

But first let's fetch them from prop's since we are putting them through the props.

So instead of using this props on the next click we can just structure them from props and get to them

and assign them as a click method.

So let's see if this works we press next and went to page 2 now Page free.

Now check out our case.

So where are you when you are on the last page you are not able to move to the next page anymore.

And with the first page the same.

So when you press previous you were staying on the first page and also clicking the individual page

works perfectly.