202.1. LoadMore component - fetching next comments (different way of pagination)(Section 21: React Redux Application)
In this lecture we'll handle another case of pagen eating it up at this time instead of setting the

page will load the More button for our comments.

So let's first couple of comments so it will have more than 10 because that's when it breaks this comments

into individual pages because that's what we have said inside our API platform.

So let's maybe one more so just find a post that already has a couple of comments and some more so you

reach more than 10.

And the thing is that you will see all those comments on this page because they are at it and we have

handled visually displaying and in your comments.

Now it looks great.

You had like 15 comments but when you refresh it's unfortunately been cut off.

So you only see the comments from the first page and there are already two pages.

You can see that there is also Hydra total items parameter.

It says there are 15 and we are showing only 10 so we need some kind of pagination but we wouldn't use

normal pugging a..

Usually with comments you have but on with the load more labels.

So let's do something like that.

It's like on Facebook you would like to load more posts and just scroll through them.

You wouldn't like to go through pages so it doesn't work for things like comments.

So let's quickly add a lot more comment component inside the load more G-S.

So as always we import react and extend the reactor component and the render method we drew in return.

Simple Day 6 which would be a button with a label that will be passed through prop's so it will just

render this label on a block button.

Now let's add this load more component inside our comment list container.

Be sure to import this component in the import statement and maybe specify the label as load more comments.

So that's how it looks like.

It's pretty big and obvious what it will do so well it doesn't do anything yet.

So now we'll have to somehow implement this functionality and seems it is a component that will work

similar to the paginate or component.

It would be a dump component so it won't care what actually gets dispatched.